requirements.txt
aiogram==3.26.0
aiogram-calendar==0.6.0
alembic==1.18.4
apscheduler==3.11.2
asyncpg==0.31.0
psycopg2-binary==2.9.11
pydantic==2.12.5
pydantic-settings==2.13.1
redis==7.3.0
sqlalchemy==2.0.48
yandexgpt-python==0.2.3